requirements.txt
# Core data science
numpy
pandas
scipy

# Visualisation
matplotlib
seaborn
missingno
ipywidgets

# Jupyter
jupyterlab
ipython

# HuggingFace / datasets
datasets<4.0.0
transformers
pyarrow

# Audio processing
soundfile
librosa
openai-whisper
jiwer

# NLP / bias benchmarks
# (crows_pairs loaded via datasets; stereoset via direct parquet URL)

# Utilities
requests
tqdm

# API clients
openai
python-dotenv
